Working signature provider
