+++ src/language/index.ts
@@ -1,12 +1,14 @@
 import { completionProvider } from "./providers/completionProvider";
 import { formatProvider } from "./providers/formatProvider";
+import { signatureProvider } from "./providers/parameterProvider";
 
 export function languageInit() {
   let functionality = [];
 
   functionality.push(
     completionProvider,
-    formatProvider
+    formatProvider,
+    signatureProvider
   );
   
   return functionality;
